Move the database functions to the worker store
