+++ synapse/storage/data_stores/main/events.py
@@ -40,12 +40,11 @@ from synapse.metrics.background_process_metrics import run_as_background_process
 from synapse.storage._base import make_in_list_sql_clause
 from synapse.storage.background_updates import BackgroundUpdateStore
 from synapse.storage.data_stores.main.event_federation import EventFederationStore
-from synapse.storage.data_stores.main.events_worker import EventsWorkerStore
+from synapse.storage.data_stores.main.events_worker import encode_json, EventsWorkerStore
 from synapse.storage.data_stores.main.state import StateGroupWorkerStore
 from synapse.types import RoomStreamToken, get_domain_from_id
 from synapse.util import batch_iter
 from synapse.util.caches.descriptors import cached, cachedInlineCallbacks
-from synapse.util.frozenutils import frozendict_json_encoder
 
 logger = logging.getLogger(__name__)
 
@@ -55,16 +54,6 @@ event_counter = Counter(
     "",
     ["type", "origin_type", "origin_entity"],
 )
-
-
-def encode_json(json_object):
-    """
-    Encode a Python object as JSON and return it in a Unicode string.
-    """
-    out = frozendict_json_encoder.encode(json_object)
-    if isinstance(out, bytes):
-        out = out.decode("utf8")
-    return out
 
 
 _EventCacheEntry = namedtuple("_EventCacheEntry", ("event", "redacted_event"))
@@ -1112,22 +1101,6 @@ class EventsStore(
 
         yield self.runInteraction("_update_censor_txn", _update_censor_txn)
 
-    def _censor_event_txn(self, txn, event_id, pruned_json):
-        """Censor an event by replacing its JSON in the event_json table with the
-        provided pruned JSON.
-
-        Args:
-            txn (LoggingTransaction): The database transaction.
-            event_id (str): The ID of the event to censor.
-            pruned_json (str): The pruned JSON
-        """
-        self._simple_update_one_txn(
-            txn,
-            table="event_json",
-            keyvalues={"event_id": event_id},
-            updatevalues={"json": pruned_json},
-        )
-
     @defer.inlineCallbacks
     def count_daily_messages(self):
         """
@@ -1966,104 +1939,6 @@ class EventsStore(
                 }
                 for label in labels
             ],
-        )
-
-    def insert_event_expiry(self, event_id, expiry_ts):
-        """Save the expiry timestamp associated with a given event ID.
-
-        Args:
-            event_id (str): The event ID the expiry timestamp is associated with.
-            expiry_ts (int): The timestamp at which to expire (delete) the event.
-        """
-        return self._simple_insert(
-            table="event_expiry",
-            values={"event_id": event_id, "expiry_ts": expiry_ts},
-            desc="insert_event_expiry",
-        )
-
-    def expire_event(self, event):
-        """Redact an event that has expired, and delete its associated expiry timestamp.
-
-        Args:
-             event (events.EventBase): The event to delete.
-        """
-        # Prune the event's dict then convert it to JSON.
-        pruned_json = encode_json(prune_event_dict(event.get_dict()))
-
-        def delete_expired_event_txn(txn):
-            # Update the event_json table to replace the event's JSON with the pruned
-            # JSON.
-            self._censor_event_txn(txn, event.event_id, pruned_json)
-
-            # Delete the expiry timestamp associated with this event from the database.
-            self._simple_delete_txn(
-                txn, table="event_expiry", keyvalues={"event_id": event.event_id}
-            )
-
-            # We need to invalidate the event cache entry for this event because we
-            # changed its content in the database.
-            txn.call_after(self._get_event_cache.invalidate, (event.event_id,))
-            # Send that invalidation to replication so that other workers also invalidate
-            # the event cache.
-            self._send_invalidation_to_replication(
-                txn, "_get_event_cache", (event.event_id,)
-            )
-
-        return self.runInteraction("delete_expired_event", delete_expired_event_txn)
-
-    def delete_event_expiry(self, event_id):
-        """Delete the expiry timestamp associated with an event ID without deleting the
-        actual event.
-
-        Args:
-            event_id (str): The event ID to delete the associated expiry timestamp of.
-        """
-        return self._simple_delete(
-            table="event_expiry",
-            keyvalues={"event_id": event_id},
-            desc="delete_event_expiry",
-        )
-
-    def get_events_to_expire(self):
-        """Retrieve the IDs of the events we have an expiry timestamp for, along with
-        said timestamp.
-
-        Returns:
-            A list of dicts, each containing an event_id and an expiry_ts.
-        """
-        return self._simple_select_list(
-            table="event_expiry",
-            keyvalues=None,
-            retcols=["event_id", "expiry_ts"],
-            desc="get_events_to_expire",
-        )
-
-    def get_next_event_to_expire(self):
-        """Retrieve the entry with the lowest expiry timestamp in the event_expiry
-        table, or None if there's no more event to expire.
-
-        Returns:
-            A dict with an event_id and an expiry_ts of there's at least one row in the
-            event_expiry table, None otherwise.
-        """
-
-        def get_next_event_to_expire_txn(txn):
-            txn.execute(
-                """
-                SELECT event_id, expiry_ts FROM event_expiry
-                ORDER BY expiry_ts ASC LIMIT 1
-                """
-            )
-
-            res = self.cursor_to_dict(txn)
-
-            if res:
-                return res[0]
-            else:
-                return None
-
-        return self.runInteraction(
-            desc="get_next_event_to_expire", func=get_next_event_to_expire_txn
         )
